Re-organizing some of the files/packages
