+++ test-automation/ui-tests/src/test/java/org/opennms/horizon/systemtests/pages/AppliancePage.java
@@ -25,7 +25,7 @@
  *     http://www.opennms.org/
  *     http://www.opennms.com/
  *******************************************************************************/
-package org.opennms.horizon.systemtests.pages.cloud;
+package org.opennms.horizon.systemtests.pages;
 
 import com.codeborne.selenide.SelenideElement;
