+++ test-automation/ui-tests/src/test/java/org/opennms/horizon/systemtests/pages/ConfirmDeleteMinionPopup.java
@@ -26,40 +26,18 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.systemtests.pages.cloud;
+package org.opennms.horizon.systemtests.pages;
 
+import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.SelenideElement;
-import org.opennms.horizon.systemtests.CucumberHooks;
-import org.openqa.selenium.By;
-import static com.codeborne.selenide.Condition.enabled;
+
 import static com.codeborne.selenide.Selenide.$;
 
-public class CloudInstanceLoginPage {
-    private static final SelenideElement usernameField = $(By.id("idp-discovery-username"));
-    private static final SelenideElement nextBtn = $(By.id("idp-discovery-submit"));
-    private static final SelenideElement passwordField = $(By.id("okta-signin-password"));
-    private static final SelenideElement submitBtn = $(By.id("okta-signin-submit"));
+public class ConfirmDeleteMinionPopup {
+    private final static SelenideElement deleteBtn = $("[data-testid='save-btn']");
 
-    public static void setUsername(String username) {
-        usernameField.shouldBe(enabled).setValue(username);
+    public static void clickDeleteBtn() {
+        deleteBtn.shouldBe(Condition.enabled).click();
     }
 
-    public static void clickNextBtn() {
-        nextBtn.shouldBe(enabled).click();
-    }
-
-    public static void setPassword(String password) {
-        passwordField.shouldBe(enabled).setValue(password);
-    }
-
-    public static void clickSubmitBtn() {
-        submitBtn.shouldBe(enabled).click();
-    }
-
-    public static void login() {
-        setUsername(CucumberHooks.adminUsername);
-        clickNextBtn();
-        setPassword(CucumberHooks.adminPassword);
-        clickSubmitBtn();
-    }
 }
