+++ test-automation/ui-tests/src/test/java/org/opennms/horizon/systemtests/steps/LeftPanelSteps.java
@@ -25,35 +25,21 @@
  *     http://www.opennms.org/
  *     http://www.opennms.com/
  *******************************************************************************/
-package org.opennms.horizon.systemtests.steps.cloud;
 
-import com.codeborne.selenide.Selenide;
-import io.cucumber.java.en.And;
+package org.opennms.horizon.systemtests.steps;
+
 import io.cucumber.java.en.Then;
-import org.opennms.horizon.systemtests.pages.cloud.CloudHomePage;
-import org.opennms.horizon.systemtests.pages.cloud.CloudLeftPanelPage;
+import org.opennms.horizon.systemtests.pages.LeftPanelPage;
 
-public class CloudHomeSteps {
-    @Then("sees {string} subtitle in the Top 10 Applications chart")
-    public void verifyErrorMessage(String subtitle) {
-        CloudLeftPanelPage.clickOnPanelSection("home");
-        CloudHomePage.verifyTop10ApplicationsSubtitle(subtitle);
+public class LeftPanelSteps {
+
+    @Then("Navigate to the {string} through the left panel")
+    public void clickOnLeftPanelSection(String section) {
+        LeftPanelPage.clickOnPanelSection(section);
     }
 
-    @Then("wait until the 'Top 10 Applications' chart will reflect the received data")
-    public void waitFlowsChart() {
-        for (int i = 0; i < 6; i++) {
-            Selenide.sleep(5_000);
-            Selenide.refresh();
-            if (CloudHomePage.verifyTop10Applications()) {
-                break;
-            }
-        }
-        CloudHomePage.verifyNoDataTop10ApplicationsState(false);
-    }
-
-    @Then("click on 'Flows' link")
-    public void clickOnFlows() {
-        CloudHomePage.clickOnFlowsLink();
+    @Then("user sees the navigation panel for instance")
+    public void verifyLeftPanel() {
+        LeftPanelPage.verifyLeftPanelIsDisplayed();
     }
 }
