+++ test-automation/ui-tests/src/test/java/org/opennms/horizon/systemtests/steps/UdpGenSteps.java
@@ -26,15 +26,30 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.systemtests.steps.cloud;
+package org.opennms.horizon.systemtests.steps;
 
 import io.cucumber.java.en.Then;
-import org.opennms.horizon.systemtests.pages.cloud.CloudAuthErrorPage;
+import lombok.SneakyThrows;
+import org.opennms.horizon.systemtests.CucumberHooks;
+import testcontainers.UpdGenContainer;
 
-public class CloudAuthErrorSteps {
-    @Then("see 'We are sorry...' error with access restriction for {string} user")
-    public void verifyErrorMessage(String email) {
-        CloudAuthErrorPage.verifyAuthError(email);
-        CloudAuthErrorPage.logout();
+import java.util.Arrays;
+
+public class UdpGenSteps {
+    @SneakyThrows
+    @Then("send {int} packets of {string} traffic to {int} port")
+    public void generateNetflow9(Integer number, String flowType, Integer port) {
+        if (!Arrays.asList("netflow9", "netflow5", "ipfix").contains(flowType)) {
+            throw new RuntimeException("Do not support this type");
+        }
+
+        try (
+            UpdGenContainer updGenContainer = new UpdGenContainer(
+                CucumberHooks.MINIONS.get(0).getUdpPortBinding(port),
+                flowType,
+                number)
+        ) {
+            updGenContainer.start();
+        }
     }
 }
