+++ test-automation/ui-tests/src/test/resources/features/appliance-tests.feature
