+++ test-automation/ui-tests/src/test/resources/features/discovery-tests.feature
